Generation Model Deployment/requirements.txt
flask
gunicorn
tensorflow
pillow
transformers==4.28.0
fake_useragent
numpy